requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pypdf2
chromadb
faiss-cpu
langchain-google-genai 

-e.